function/requirements.txt
google-cloud-dns==0.35.0
dnspython